+++ actix-web/CHANGES.md
@@ -9,6 +9,7 @@
 - Add `web::Payload::to_bytes[_limited]()` helper methods.
 - Add missing constructors on `HttpResponse` for several status codes.
 - Add `http::header::ContentLength` typed header.
+- Add `HttpRequest::full_uri()` method to get the full uri of an incoming request.
 
 ### Changed